# docs/swagger.yaml
definitions:
  auth.User:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.User'
      responses:
        "200":
          description: OK
      summary: Register a new user
      tags:
      - auth
  /api/auth/token:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get auth token for the user
      tags:
      - auth
...
  /api/images/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      - description: Auth Header
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Upload an image
      tags:
      - images
...
swagger: "2.0"